common: Fix return type of march/mimpid helpers

+++ include/common/jumpstart.h
@@ -97,8 +97,8 @@ uint64_t get_thread_attributes_trap_override_struct_address_from_smode(void);
 uint8_t get_thread_attributes_current_mode_from_smode(void);
 uint8_t get_thread_attributes_current_v_bit_from_smode(void);
 uint8_t get_thread_attributes_hart_id_from_smode(void);
-uint8_t get_thread_attributes_marchid_from_smode(void);
-uint8_t get_thread_attributes_mimpid_from_smode(void);
+uint64_t get_thread_attributes_marchid_from_smode(void);
+uint64_t get_thread_attributes_mimpid_from_smode(void);
 uint8_t get_thread_attributes_vsmode_setup_done_from_smode(void);
 uint8_t
 get_thread_attributes_num_context_saves_remaining_in_smode_from_smode(void);
@@ -110,8 +110,8 @@ uint64_t get_thread_attributes_trap_override_struct_address_from_mmode(void);
 uint8_t get_thread_attributes_current_mode_from_mmode(void);
 uint8_t get_thread_attributes_current_v_bit_from_mmode(void);
 uint8_t get_thread_attributes_hart_id_from_mmode(void);
-uint8_t get_thread_attributes_marchid_from_mmode(void);
-uint8_t get_thread_attributes_mimpid_from_mmode(void);
+uint64_t get_thread_attributes_marchid_from_mmode(void);
+uint64_t get_thread_attributes_mimpid_from_mmode(void);
 uint8_t get_thread_attributes_smode_setup_done_from_mmode(void);
 uint8_t
 get_thread_attributes_num_context_saves_remaining_in_mmode_from_mmode(void);
